Updates @typescript-eslint/* major version and other minor updates
As part of updating the major version from 2.X to 3.X, I cleaned up
the rule overrides in .eslintrc.js. I added `prefer-const: 0` because
the new `plugin:@typescript-eslint/recommended` ruleset sets
`prefer-const: 2`, which results in a bunch of transformations of the
ts and js templates. I'll unset that rule and re-run in a subsequent
commit.